// src/types/ao-kpi.ts
export interface AoKpiExcelRow {
  TanggalTarik?: string | number | Date;
  CabangId?: string | number;
  AOId?: string | number;
  NamaAO?: string;
  Jumlah_NOC?: number;
  Jumlah_NOA?: number;
  OS?: number;
  Jumlah_LAR?: number;
  Jumlah_Kelompok?: number;
  Penambabahan_Lar_Baru?: number;
  Penambabahan_OS_Lar_Baru?: number;
  OS_1A_EOM?: number;
  Target_Lar_Baru?: number;
  Realisasi_Persen_Lar_Baru?: number;
  Pencapaian_Lar_Baru?: number;
  Bobot_Pencapaian_Lar_Baru?: number;
  Nilai_Pencapaian_Lar_Baru?: number;
  Target_Ao_S1_Harian?: number;
  Target_Ao_SL_Harian?: number;
  Realisasi_S1?: number;
  Realisasi_SL?: number;
  Pencapaian_UK_S1?: number;
  Bobot_UK_S1?: number;
  Nilai_UK_S1?: number;
  Pencapaian_UK_SL?: number;
  Bobot_UK_SL?: number;
  Nilai_UK_SL?: number;
  Persen_Hadir_BayarFullPayment?: number;
  Target_Persen_Hadir_BayarFullPayment?: number;
  Pencapaian_Hadir_BayarFullPayment?: number;
  Bobot_Hadir_BayarFullPayment?: number;
  Nilai_Hadir_BayarFullPayment?: number;
  Pencairan_S1?: number;
  Pencairan_SL?: number;
  Persen_Pencairan_S1?: number;
  Persen_Pencairan_SL?: number;
  Qualifier_S1?: string;
  Qualifier_SL?: string;
  Pencapaian_UK_S1_New?: number;
  Pencapaian_UK_SL_New?: number;
  Total_Nilai?: number;
}